+++ lib/Admin/Admin.php
@@ -107,8 +107,6 @@ class Admin {
 		<?php
 	}
 
-
-
 	/**
 	 * Enqueues admin scripts.
 	 *